% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikC}
\alias{LogLikC}
\title{Employ the Kalman Filter to calculate the loglikelihood}
\usage{
LogLikC(y, y_isna, a, P_inf, P_star, Z, T, R, Q)
}
\arguments{
\item{y}{Matrix of observations.}

\item{y_isna}{Matrix indicating which observations are missing.}

\item{a}{Initial values of the state vector.}

\item{P_inf}{Diffuse part of the variance - covariance matrix of the
state vector.}

\item{P_star}{Stationary part of the variance - covariance matrix of the
state vector.}

\item{Z}{Z system matrix of the State Space model.}

\item{T}{T system matrix of the State Space model.}

\item{R}{R system matrix of the State Space model.}

\item{Q}{Q system matrix of the State Space model.}
}
\description{
Employ the Kalman Filter to calculate the loglikelihood
}
