% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StateSpaceEval.R
\name{StateSpaceEval}
\alias{StateSpaceEval}
\title{State Space Model Evaluation at Specified Parameters}
\usage{
StateSpaceEval(
  param,
  y,
  H_format = NULL,
  local_level_ind = FALSE,
  slope_ind = FALSE,
  BSM_vec = NULL,
  cycle_ind = FALSE,
  addvar_list = NULL,
  level_addvar_list = NULL,
  slope_addvar_list = NULL,
  exclude_level = NULL,
  exclude_slope = NULL,
  exclude_BSM_list = lapply(BSM_vec, FUN = function(x) {     0 }),
  exclude_cycle_list = list(0),
  format_level = NULL,
  format_slope = NULL,
  format_BSM_list = lapply(BSM_vec, FUN = function(x) {     NULL }),
  format_cycle_list = list(NULL),
  format_addvar = NULL,
  format_level_addvar = NULL,
  loglik_only = FALSE
)
}
\arguments{
\item{param}{Parameters used to construct the system matrices.}

\item{y}{N x p matrix containing the N observations of the p dependent variables.}

\item{H_format}{Format of the H system matrix,
the variance - covariance matrix of the observation equation.}

\item{local_level_ind}{Boolean indicating whether a local level should be added
to the state space model.}

\item{slope_ind}{Boolean indicating whether a local level + slope should be added
to the state space model.}

\item{BSM_vec}{Vector containing the BSM seasonalities that have to be added
to the state space model.}

\item{cycle_ind}{Boolean indicating whether a cycle has to be added
to the state space model.}

\item{addvar_list}{A list containing the explanatory variables for each of the
dependent variables. The list should contain p (number of dependent variables)
elements. Each element of the list should be a N x k_p matrix,
with k_p being the number of explanatory variables for the pth dependent variable.
If no explanatory variables should be added for one of the dependent variables,
then set the corresponding element to \code{NULL}.}

\item{level_addvar_list}{A list containing the explanatory variables for each of
the dependent variables. The list should contain p (number of dependent variables)
elements. Each element of the list should be a N x k_p matrix,
with k_p being the number of explanatory variables for the pth dependent variable.
If no explanatory variables should be added for one of the dependent variables,
then set the corresponding element to \code{NULL}.}

\item{slope_addvar_list}{A list containing the explanatory variables for each of
the dependent variables. The list should contain p (number of dependent variables)
elements. Each element of the list should be a N x k_p matrix,
with k_p being the number of explanatory variables for the pth dependent variable.
If no explanatory variables should be added for one of the dependent variables,
then set the corresponding element to \code{NULL}.}

\item{exclude_level}{Vector containing the dependent variables that should not
get a local level.}

\item{exclude_slope}{Vector containing the dependent variables that should not
get a slope.}

\item{exclude_BSM_list}{List of vectors, each vector containing the dependent variables
that should not get the corresponding BSM component.}

\item{exclude_cycle_list}{The dependent variables that should not get a cycle component.
Should be a list to allow different dependent variables to be excluded for
different cycles.}

\item{format_level}{Format of the Q_level system matrix the variance - covariance matrix
of the level state equation.}

\item{format_slope}{Format of the Q_slope system matrix, the variance - covariance matrix
of the slope state equation.}

\item{format_BSM_list}{Format of the Q_BSM system matrix, the variance - covariance matrix
of the BSM state equation. Should be a list to allow different formats for different
seasonality periods.}

\item{format_cycle_list}{Format of the Q_cycle system matrix, the variance - covariance
matrix of the cycle state equation. Should be a list to allow different formats
for different cycles.}

\item{format_addvar}{Format of the Q_addvar system matrix, the variance - covariance matrix
of the explanatory variables state equation.}

\item{format_level_addvar}{Format of the Q_level_addvar system matrix, the
variance - covariance matrix of the explanatory variables of the level state equation.}

\item{loglik_only}{Boolean indicating whether only the loglikelihood should be returned.}
}
\value{
A list containing:
\itemize{
\item function_call: A list containing the input to the function.
\item system_matrices: A list containing the system matrices of the State Space model.
\item predicted: A list containing the predicted components of the State Space model.
\item filtered: A list containing the filtered components of the State Space model.
\item smoothed: A list containing the smoothed components of the State Space model.
\item diagnostics: A list containing items useful for diagnostical tests.
}
}
\description{
Evaluates the specified State Space model at the parameters specified by the user.
}
\examples{
# Evaluates a local level model for the Nile data
y <- matrix(Nile)
StateSpaceEval(param = c(1, 1), y = y / 100, local_level_ind = TRUE)

}
\references{
\insertRef{durbin2012time}{statespacer}
}
\author{
Dylan Beijers, \email{dylanbeijers@gmail.com}
}
